# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI TimeShards API",
    "version": "0.1.0",
    "description": "REST API for time tracking and access control. Narrative reference: docs/API.md. Machine-readable spec served at GET /api/v1/openapi.json."
  },
  "servers": [{ "url": "http://127.0.0.1:47821", "description": "Default local bind" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token from POST /api/v1/auth/login"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": { "message": { "type": "string" } }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string", "format": "password" }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "security": [],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Bearer token and user summary" },
          "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "Default password blocked (production/staging env)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  },
  "security": [{ "bearerAuth": [] }]
}